This change heads off an issue that is soon to be made real by the inclusion of ImageBundler. In general, the change allows Generators (in the compiler, hosted mode, or in JUnit tests) to make AWT dependent calls without hanging the application on OS X.
GWTCompiler and GWTShell have been changed to ensure that we always
exit through a call to System.exit in the static main. GWTCompiler had
to be refactored to change the threading for the case when the
-treeLogger flag is passed in. Specificaly, the compiler now forks a
thread and runs the main compile on the secondary thread to allow the
main thread to pump the SWT event loop associated with the tree logger
window. A system property has also been set in BootStrapPlatform to
disable CocoaComponent compatibility mode for OS X.

Review by: scottb
           bruce



git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@658 8db76d5a-ed1c-0410-87a9-c151d255dfc7