Fixes issue #1260. I introduced the bug in r955. The problem I was trying to solve in that revision was making the following code do an array store check:
FinalType ft[];
NonFinalSuper nfs;
ft[0] = nfs;

Previously, we'd fail to do a type check on account of the element type being final; but really you need the element type to be final AND the RHS static type to be the same type to elide the check.

My faulty logic caused a set to be emitted any time the element type and RHS static type differed-- even if it was a primitive array.  This change makes sure primitive arrays never emit a type check.

Found by: ispeters
Review by: mmendez


git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@1198 8db76d5a-ed1c-0410-87a9-c151d255dfc7