There is a problem when using Internet Explorer to access image bundle files over HTTPS, when the image bundle files have a security constraint set by the web application. The problem seems to occur when the web application server is Tomcat or Glassfish.  It stems from the way in which Internet Explorer renders the composite image, and the caching headers that are set by Tomcat and Glassfish when the HTTPS protocol is used to request a resource protected by a security constraint.
Tomcat and Glassfish set HTTP response headers indicating that the file should
not be cached. Internet Explorer requires that all content viewed via a plugin
must be cached. We are using AlphaImageLoader to display png files, which
qualifies as a plugin. The result is that the images which rely on the image
bundle image are not displayed. 

After doing some research, I discovered a few ways to get around this problem.
Since this problem is not really GWT specific, the best that I could do is
document the problem and the workarounds in the ImageBundle javadoc. I have also
added a "Tip" to the ImageBundle section of the developer guide to point users
to the ImageBundle javadoc if they plan on using HTTPS with ImageBundles. 

I also added another section to the ImageBundle javadoc about caching image
bundle files. This section was already in the Developer Guide, but it is good
have it in both places, as not everybody reads the Developer Guide. 

Finally, I made a couple minor corrections to the existing javadoc.

Patch by: rdayal
Review by: jgw
Issue: 1172



git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@1200 8db76d5a-ed1c-0410-87a9-c151d255dfc7