Now that PopupPanel.setWidth(String) and PopupPanel.setHeight(String) actually set the dimensions of the PopupPanel's child widget instead of the PopupPanel itself, there are behavioral differences between PopupPanel.getOffsetHeight() and PopupPanel.getOffsetWidth() between GWT 1.3 and GWT 1.4RC.
PopupPanel.setWidth/PopupPanel.setHeight will only take effect if the panel has
a child widget set. If PopupPanel.setWidth /PopupPanel.setHeight is called
before a child widget for the panel has been set, then
PopupPanel.getOffsetWidth/PopupPanel.getOffsetHeight will NOT include the
width/height that was set. 

In the same scenario in GWT 1.3, the width/height that was set would be
included in the computation of getOffsetWidth/getOffsetHeight, regardless of
whether or not a child widget was set. 

To clarify this difference, javadoc has been added to the setWidth, setHeight,
getOffsetWidth, and getOffsetHeight methods.

Issue: 1148
Patch by: rdayal
Review by: jgw



git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@1222 8db76d5a-ed1c-0410-87a9-c151d255dfc7