When removing a widget from the RootPanel, the RootPanel overwrites the CSS positioning attributes of the Widget in the changeToStaticPositioning method. The idea was to revert any changes made by the Panel, but it actually overwrites the position attribute from a CSS file instead.
This bug also affects the PopupPanel, which uses the left and top attributes to
position itself.  The changeToSTaticPositioning method overwrites the left and
top attributes, effectively repositioning the popup every time it is hidden.

Fix: Instead of setting the position attribute to "static", I set it to an
empty string, reverting back to the original CSS attributes.  I added a couple
of variables to the PopupPanel to save the top and left position as they are
set.  When the PopupPanel is shown, it is automatically repositioned to the
previous location.

Unit Test: I modified PopupTest to take the PopupPanel changes into account by
hiding a PopupPanel, showing it, and verifying it is still in the correct
location.

Issue: 1195
Patch by: jlabanca
Review by: jgw



git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@1244 8db76d5a-ed1c-0410-87a9-c151d255dfc7