Description: ColumnFormatter.setWidth changes the width of the wrong column index in Mozilla, but works correctly for all other browsers.  The issue only occurs when setColumnWidth is called after the HTMLTable widget is added to the page.  It appears that Mozilla creates an internal reference to a colgroup if none exists when the Element is added to the page, shifting the index of all the columns by one.  This bug affects the FlexTable and Grid classes.
Fix:
I added a call to prepareColumnGroup in the setColumnFormatter method,
which guarantees that we create a col tag when the HTMLTable is instantiated
(before adding it to the page).  This also means that we have a col tag even
when we have no columns, but this does not appear to affect anything.

Issue: 1223
Patch by: jlabanca
Review by: jgw



git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@1283 8db76d5a-ed1c-0410-87a9-c151d255dfc7