The RadioButton.setName changes the RadioButton's group name.  In fact, radio inputs are actually grouped by their name in pure HTML, so this is the intended behavoir.  However, the usage of the term "group" in the constructor and "name" is a little misleading, as they are the same attribute.
Fix:
I modified all of the RadioButton specific methods to use the term "name"
instead of "group", and I added more JavaDoc comments explaining the
relationship between name and group.  There is no change in logic in this fix,
just some rewording and clearer variable names.

In RadioButton.setName(), I remove the old inputElement and replace it with a
new identical one.  I transfer all of the attributes of the old element to the
new one, such as enabled, checked, accessKey, and uid.  By setting the checked
status of the new element, we can guarantee we will never have two checked
radio buttons in the same group.

I tested this change on every browser and it worked without error.  I also
added the radios to a form to verify that the radio is associated with its new
group on form submittal.  Some of the things I looked for are: 

- User should not see a flicker when changing element
- RadioButton methods should affect the new input element
- New input element should be in the same state as the old element
- No more than one item should be selected in each of the old and new radio button groups 
- AccessKey and tabIndex should now apply to the new element
- The submitted form data should include the radio button in the new group

Issue: 1262
Patch by: jlabanca
Review by: jgw



git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@1292 8db76d5a-ed1c-0410-87a9-c151d255dfc7