This patch addresses the following problems with the TypeOracle:
JParamterizedTypes we reporting too many subtypes if the JParameterizedType was constructed before its corresponding generic type was completely contructed.  Changed to lazy initialization of the substitution map.  Also updated the isAssignableFrom and getSubtype logic.

TypeOracle was throwning an NPE if gwt.typeArgs were applied to a type that was not the raw version of a generic type.  This has been changed to an error and logged accordingly.

TypeOracle used to report all gwt.typeArgs parsing errors as generic could-not-find-type error messages even though more specific failure information was avaiable.  The more detailed error messages are now logged to the shell.

Patch by: mmendez
Review by: scottb (desk check)

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@1543 8db76d5a-ed1c-0410-87a9-c151d255dfc7