Fixes "Unresolvable native reference to method 'instantiate' in type {TypeName}_FieldSerializer" errors introduced in the GWT 1.5 code base.
Generated field serializers should always define an instantiate method for a class whether or not it maybe instantiated by the RPC, client-side proxy.  This is required because field serializers are shared by all of the RPC, client-side proxies in a compilation and these proxies may have conflicting views of a type's instantiability.

Review by: Bobv


git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@1699 8db76d5a-ed1c-0410-87a9-c151d255dfc7