Fixes issue 1830.  Modified AbstractCompiler.INameEnvironmentImpl.findType(char[][]) to ask the context class loader for a type the binary version of a class if there is no source code for it.  This allows JDT compiles to complete normally.
The TypeOracle was updated to ingnore any annotations whose type declaration was only available in binary form; the compiler always ignored annotations so no modifications were required for it.

The TypeOracle and the compiler will generate warnings and errors, respectively, if a binary only class is used outside the context of an annotation.

Patch by: mmendez
Review by: scottb

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@1748 8db76d5a-ed1c-0410-87a9-c151d255dfc7