Fixes the problem reported in issue 1830 comment #12.
We were always telling JDT that any binary only package was not a package.  This would work as long as JDT always called INameEnvironment.findType for a class in a binary only package.  

This patch uses ClassLoader.getResource(String) to determine if a name is a package.  This works as long as we are not asking about binary only package which only exists on the boot classpath.

Review by: scottb

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@1944 8db76d5a-ed1c-0410-87a9-c151d255dfc7