New APIs in RequestBuilder to add the following three methods:
void setRequestData(String)
void setCallback(RequestCallback)
Request send()

Calling each of these methods in turn will have a similar effect to calling the existing method:

sendRequest(String, RequestCallback)

The main difference is that the additional setters allow the data to be stored in the RequestBuilder object.  This change will enable RPC to return a "ready to go" RequestBuilder for a client to tweak.

Review by: bobv


git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@2243 8db76d5a-ed1c-0410-87a9-c151d255dfc7