Fixes issue 1695
Adds the command line argumetns -addToClassPath and -addModule 
(for applicationCreator) to:

  - applicationCreator
  - projectCreator
  - junitCreator

To allow an external library to be added as a part of the initial specification 
of a project.  It does some checks to see that the library exists and that 
the module exists somewhere within the classpath.

Suggested by: mmendez
Patch by: zundel
Review by: jat (desk check)



git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@2904 8db76d5a-ed1c-0410-87a9-c151d255dfc7