DatePicker throws an IllegalStateException when it is used in a time zone where daylight savings time occurs at midnight.  The problem is that, as the DatePicker populates the dates in the calendar, it uses a time of 00:00:00 for all dates.  However, in countries where DST occurs at midnight, the dates are coerced to the next day, resulting in an invalid state.
With this patch, DatePicker uses a time of noon instead of midnight. DST always occurs between midnight and 5am, so noon is safe.  This patch also fixes a broken test that uses date.toString() instead of formatting the date for the DateBox being testing.

Patch by: jlabanca
Review by: rjrjr



git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@5218 8db76d5a-ed1c-0410-87a9-c151d255dfc7