Changes the code splitter to use a simpler splitting strategy. There is now only a single leftovers fragment.  Additionally, each split point gets either a base or an exclusive fragment. As a result, all code ends up in exactly one fragment.
An additional trick in the new splitting strategy is that
the splitter tries to predict which split point will be
reached first.  If there is such a one, it is given a base
fragment instead of an exclusive, and it can be loaded
without waiting for the leftovers.  This process is then
repeated, trying to find a split point definitely reached
second, etc., until no such split point is found.

Review by: scottb



git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@5293 8db76d5a-ed1c-0410-87a9-c151d255dfc7