During precompile, only do a draft optimize due to superlinear optimization performance.
This makes precompile faster at the expense of each sharded compile being slower.  This is a win if youc can properly parallelize the individual compile shards, but we should revisit this change for users who only have a single build machine, as it could be a net loss for them.

This also makes draft compiles (as a feature) faster by doing even fewer optimizations.

Review by: spoon

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@5338 8db76d5a-ed1c-0410-87a9-c151d255dfc7