Big refactoring of StandardGeneratorContext so that only one instance is created during a compile.
Many things now share the same StandardGeneratorContext, which allows .finish() calls to be chunked, which is somewhat faster in some cases.  Also, a lot of plumbing gets cleaner because we avoid situations where we had to pass around all the components needed to create a new StandardGeneratorContext on the fly.

Review by: spoon

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@5340 8db76d5a-ed1c-0410-87a9-c151d255dfc7