Optimize the output production by reusing 16k buffers over and over.
By reusing over and over a buffer small enough to fit into the processor's L2 cache, we realize a huge savings on data transfer.  This replaces the previous method of reading data entirely into huge byte[] and then writing it back in one shot.  Experimentally, this is shown to be much faster.

Review by: spoon, fabbott

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@5404 8db76d5a-ed1c-0410-87a9-c151d255dfc7