Implement link-to-zip archive, which can be faster for big links.
When a user specifies a -war or -extra param that ends with ".zip", ".jar", or ".war", produce output as an archive instead of a filesystem folder.  This feature is intended for build systems which Link, then immediately zip up the output directory.  Doing this within the Link phase, is actually much faster, since it bypasses the need to create a whole directory tree on disk.

Still TODO: figure out a way to parallelize the compression across multiple processors.

Review by: spoon, fabbott

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@5406 8db76d5a-ed1c-0410-87a9-c151d255dfc7