Nukes the deprecated and over complicated create-the-root-now-create-the-rest-later flow from UiBinder, and in the process squashes a number of reported and unreported bugs.
Not as big a change as it looks, most of the listed files are a result of
compulsive method name change refactoring.

The real meat is in moving the parsing of the root UiBinder element out of
UiBinderWriter and into its own parser, and greatly simplified bootstrapping of
the parsing process.

Because we now start parsing above the UI, at the ui:UiBinder element,
parseElementToField (was parseWidget) gets a crack at the topmost element of the
UI--it gets treated like every other element, and various inconsistencies in its
handling (esp. with {computed.attributes} and localizable ones) evaporate.

Reviewed by: jgw


git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@6062 8db76d5a-ed1c-0410-87a9-c151d255dfc7