All the hard parts of http://code.google.com/p/google-web-toolkit/issues/detail?id=3984 are done.
We will now generate a CssResource interface for you, though you still have to
bring in the actual css from a separate .css file. The next step is to allow css
inside the ui.xml file, but that'll be easy now that all the groundwork is
done.

Note that changes were required to CssResourceGenerator to allow it to handle
generated CssResource interfaces.

Also refactored instantiating PrintWriters and logging--no more redundant copies
of die() and warn(), and no more invisible errors logged to stderr instead of
the TreeLogger.

This cleanup also lead me to move XML knowledge out of UiBinderGenerator and
into UiBinderWriter, but that was a completely mechanical
refactor--parseXmlResource has not changed.

reviewed by: spoon


git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@6102 8db76d5a-ed1c-0410-87a9-c151d255dfc7