Forgetting a namespace prefix on a child elem of DockLayoutPanel causes an NPE, for example:
<g:DockLayoutPanel unit="PCT">
 <west><g:HTMLPanel>foo</g:HTMLPanel></west>
 <east></east>
</g:DockLayoutPanel>

Unfortunately, even after the patch the error message is confusing. It
says:
"In <g:DockLayoutPanel unit='PCT'>, child must be one of {north, south,
east, west, center}"

The problem is that when you look at the XML, it seems like you've done
exactly what's being asked for. The error message ought to someone
account for the NS mismatch in how it explains the error.



git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@6643 8db76d5a-ed1c-0410-87a9-c151d255dfc7