Hello DialogBoxParser, and stricter parsing
DialogBox is weird--it implements both HasHTML and HasWidgets. This confuses the
bejeebers out of UiBinder and leads it to generate bad code, so a custom parser
is required. I took this excuse to generalize the mock UiBinderWriter work I did
yesterday for DockLayoutPanelParserTest.

I also found some low hanging fruit in XMLElement to allow it to fail when there
are stray bits of text or widgets in places they shouldn't be. Previously this
would result in quietly dropped text, or in things like <gwt:Label> showing up
the rendered DOM, sometimes even user visible.

Review by bobv
Fixes issues 4202

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@6774 8db76d5a-ed1c-0410-87a9-c151d255dfc7