{Field.references} are now validated. Also fixes unreported code gen bug when size attributes were missing under g:DockLayoutPanel
All AttributeParser implementations register any field references they encounter
with the FieldManager. When parsing is done, all references are validated for
the existence of the fields and the methods they refer to, and for correct
return types. Multiple validation errors can be reported.

The real meat of the patch is kicked off at FieldManager#validate, into
FieldReference#validate, and through the two implementations of
FieldWriter#getReturnType in AbstractFieldWriter and
FieldWriterOfGeneratedCssResource

This required retooling the AttributeParser world, so there is a lot of
collateral change across that package. Their dependencies are now provided via
constructor injection, and they are no longer instantiated reflectively (thank
goodness). The dependency chain is actually something like AttributeParser ->
FieldReferenceConverter -> FieldManager.

Custom ElementParsers must now be more disciplined, and provide the JType they
intend a consumed attribute to provide. (Most did this already.)

Reviewed by jgw, jat
http://gwt-code-reviews.appspot.com/102804

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@6867 8db76d5a-ed1c-0410-87a9-c151d255dfc7