UiBinder AttributeParsers can now distinguish "" from null
When I revamped AttributeParsing last week, I made it impossible to set
styleName to an empty string. That's a regression.

Attribute parsing has inconsistently made the assumption that setting an
attribute to "" is the same as not setting it. That is, <foo bar=""> == <foo />.
Following that rule during the revamp is part of what lead to this bug--turns
out that people already rely on <foo bar=''> causing foo.bar("") to be
generated.

So, this patch makes the parsers consistently return null for a missing
attribute, and "" only when the user actually provides it.

Reviewed by bobv
http://gwt-code-reviews.appspot.com/112812

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@7185 8db76d5a-ed1c-0410-87a9-c151d255dfc7