Fix datatype renderers to handle nulls properly for some primitive types.  We now expect renderers to produce "" for null, and parsers to return null for "".
Also works around a bug in report creation. The Created date is not be
set, and the details view cannot render a null date (because it
doesn't use renderers...). This puts us a bit out of sync with what
Roo is generating, but clears some roadblocks.

Review at http://gwt-code-reviews.appspot.com/725804


git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@8503 8db76d5a-ed1c-0410-87a9-c151d255dfc7