Changes required for reliable editing of "ui:field" attribute in template and corresponding @UiField in Java. Solved problems:
1. When GWT Designer calls createAndBindUi(), it can not provide
"owner", because we can not instantiate it, and it is not interface,
so we can not mock it too. So, we use "null" and this causes NPE
during assigning widget into fields. At design time we don't need
these fields, so null check is probably safe.

2. At design time we create ClassLoader only once and can not refresh
Java type.  So, when user asks to remove "ui:field" we update both
template and Java, but generator does not know about Java change. Same
for rename.

Patch by konstantin.scheglov@gmail.com
Review by rjrjr@google.com

http://gwt-code-reviews.appspot.com/792801/show

Review by: robertvawter@google.com

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@8675 8db76d5a-ed1c-0410-87a9-c151d255dfc7