Changes in com.google.gwt.app land to catch up to the RequestFactory overhaul. In particular, nukes PropertyColumn, and reworks AbstractProxyEdit now that create semantics have changed. Made APE a bit less monolithic in the process.
Includes a change to (and test for) RequestFactoryEditorDriver to
allow it to be initialized without access to a factory, made
after discussion with Bob.

I will still be busting up com.google.gwt.app and banishing these
proxy specific places and activities to sample land, but this is a
necessary first step.

Review at http://gwt-code-reviews.appspot.com/946802

Review by: cromwellian@google.com

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@8928 8db76d5a-ed1c-0410-87a9-c151d255dfc7