Supress errors when building the Type Oracle.
This patch supresses errors (turns them into TRACE and DEBUG level log messages)
during the TypeOracle build.  When building the AST, if an error is encountered, the
CompilationState (or global CompilationStateBuilder cache) is consulted to
find any errors in a type that has a problem, or any dependencies of that type.

Review at http://gwt-code-reviews.appspot.com/1385810


git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@9940 8db76d5a-ed1c-0410-87a9-c151d255dfc7