Adds {moz,webkit}RequestAnimationFrame support to animations.
Refactor Animation with different implementations, adding
mozRequestAnimationFrame and webkitRequestAnimationFrame support in addition to
the timer-based implementation.

ALso adds run() overloads taking an 'element' argument, that visually "scopes"
the animation (so that browsers can, as an optimization, skip steps when the
element is not visible to the user).

Code Review: http://gwt-code-reviews.appspot.com/1355805/

Author: tbroyer
Review by: jlabanca


git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@9970 8db76d5a-ed1c-0410-87a9-c151d255dfc7