Step one in making mobilewebapp more DI friendly. The goal is to remove all knowledge of ClientFactory from any class but the entry point.
This gets the basic pattern in place, submitting it as a checkpoint so
that new features can appear in the right place.

In particular, introduces the App class, and elminates a lot of low
hanging ClientFactory leakage. The interesting bit will be getting it
out of the activities. And a particularly nasty spot is where
MobileWebAppShellTablet (view class) is kicking off a
TaskListActivity. That's next.

Review at http://gwt-code-reviews.appspot.com/1433801

Review by: jlabanca@google.com

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@10163 8db76d5a-ed1c-0410-87a9-c151d255dfc7