Attachable elements in UiBinder: always run logicalAdd(Attachable) for Attachable widgets.
Before this patch, if we declared a <gwt:Attachable/> in the UiBinder template
and a corresponding @UiField(provided=true) field in the Java class, no event
handlers were attached to the widget.

This happened because of a check in UiBinder's AttachableInterpreter. It invoked
logicalAdd only if the field derived from IsWidget, which is not the case of
Attachable. As we were declaring it only as an Attachable, it did not run
logicalAdd over it, so the event handlers were not installed.

logicalAdd(Attachable) does all type checks necessary. This patch makes UiBinder always produce code to logicalAdd(Attachable) the widget declared as <gwt:Attachable/> in the template, which causes the event handlers to get attached correctly.

Review at http://gwt-code-reviews.appspot.com/1446801


git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@10182 8db76d5a-ed1c-0410-87a9-c151d255dfc7