Resolve PotentialElement children before inserting them into a container.
Before this patch, if someone tried to insertChild or insertBefore a
renderable child into a renderable container, a DOM exception was thrown, as
the element could not be found in the document. This patch follows the same
pattern as appendChild: it first tries to resolve the child being inserted to
get the real element.

While in the neighborhood, fixed some Javadoc tags related to
PotentialElement.

Review at http://gwt-code-reviews.appspot.com/1454810


git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@10315 8db76d5a-ed1c-0410-87a9-c151d255dfc7