In non-pretty mode, adds interning for castable-type-map literals. For any castmap occuring more than once, or occuring on an array type, it is hoisted to a top level variable, and it is referred to by name.
This should improve the performance of arrays and reduce uncompressed code size slightly, although it will increase the initial fragment size a little big without Closure Compiler usage. This is because all of the interned literals are added to the initial fragment, even those not used until a later fragment. With Closure, these will be moved near their point of use.

Review by: johnlenz@google.com

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@10825 8db76d5a-ed1c-0410-87a9-c151d255dfc7